backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
pydantic-settings>=2.1.0

# Redis
redis[hiredis]==5.0.1

# ML and NLP
torch>=2.0.0
transformers>=4.35.0
onnxruntime>=1.16.0
scikit-learn>=1.3.0

# URL Analysis
tldextract==5.1.1
python-whois==0.8.0

# HTTP Client
httpx==0.25.2

# LangGraph
langgraph>=0.0.26
langchain-core>=0.1.25

# Utilities
python-dotenv==1.0.0
